+++ src/Helpers/SocialMediaUpdater.php
@@ -52,10 +52,14 @@ class SocialMediaUpdater extends AbstractHelper {
 
         if ($followersNew) {
             // Update follower number and the timestamp.
-            $data['followers_updated'] = Database::getInstance()->updateFollowers($url, $followersNew);
+            if (Database::getInstance()->updateFollowers($url, $followersNew)) {
+                $data['updated'] = 'followers';
+            }
         } else {
             // Update the timestamp
-            $data['timestamp_updated'] = Database::getInstance()->updateFollowers($url, false);
+            if (Database::getInstance()->updateFollowers($url, false)) {
+                $data['updated'] = 'timestamp';
+            }
         }
 
         return $data;
@@ -67,7 +71,7 @@ class SocialMediaUpdater extends AbstractHelper {
      * @param string $network the social network
      * @param string $url the URL to check
      *
-     * @return int the follower count; negative value in case of errors.
+     * @return int|bool the follower count; false in case of errors.
      */
     private function getFollowers($network, $url) {
         switch ($network) {
@@ -75,6 +79,8 @@ class SocialMediaUpdater extends AbstractHelper {
                 return $this->getFacebookLikes($url);
             case 'googlePlus':
                 return $this->getGooglePlusOnes($url);
+            case 'instagram':
+                return $this->getInstagramFollowers($url);
             case 'twitter':
                 return $this->getTwitterFollower($url);
             case 'youtube':
@@ -89,20 +95,19 @@ class SocialMediaUpdater extends AbstractHelper {
      *
      * @param string $url the URL of the Facebook page to check
      *
-     * @return int the number of likes, or false on failure
+     * @return int|bool the number of likes, or false on failure
      */
     private function getFacebookLikes($url) {
         try {
             $id = substr($url, 25); // 25 = strlen('https://www.facebook.com/')
             if (SocialMediaUpdater::startsWith($id, 'groups/')) {
-                return 0;
+                return false;
             }
-            if ( SocialMediaUpdater::startsWith($id, 'pages/') ) {
+            if (SocialMediaUpdater::startsWith($id, 'pages/')) {
                 $temp = explode('/', $id);
                 $id = end($temp);
             }
 
-            //Construct a Facebook URL
             $json_url ='https://graph.facebook.com/' . $id .
                        '?access_token='.FACEBOOK_API_ID.'|'.FACEBOOK_API_SECRET.'&fields=likes';
             $json = @file_get_contents($json_url);
@@ -112,18 +117,15 @@ class SocialMediaUpdater extends AbstractHelper {
                 $json_url ='https://graph.facebook.com/' . $id .
                            '?access_token='.FACEBOOK_API_ID.'|'.FACEBOOK_API_SECRET.'&fields=likes';
                 $json = @file_get_contents($json_url);
-                if (!$json) {
-                    return false;
+            }
+
+            if ($json) {
+                $json = json_decode($json);
+                if (isset($json->likes)) {
+                    return $json->likes;
                 }
             }
-            $json_output = json_decode($json);
-
-            //Extract the likes count from the JSON object
-            if ($json_output->likes) {
-                return $json_output->likes;
-            } else {
-                return false;
-            }
+            return false;
         } catch (Exception $e) {
             return false;
         }
@@ -134,19 +136,42 @@ class SocialMediaUpdater extends AbstractHelper {
      *
      * @param string $url the Google+ URL to check
      *
-     * @return int the number of +1s, or false on failure
+     * @return int|bool the number of +1s, or false on failure
      */
     private function getGooglePlusOnes($url) {
         try {
             $id = substr($url, 24);
             $json_url = 'https://www.googleapis.com/plus/v1/people/' . $id .'?key=' . GOOGLE_API_KEY;
             $json = @file_get_contents($json_url);
-            if (!$json) {
-                return false;
+            if ($json) {
+                $json = json_decode($json);
+                if (isset($json->circledByCount)) {
+                    return intval($json->circledByCount);
+                }
             }
-            $json_output = json_decode($json);
+            return false;
+        } catch (Exception $e) {
+            return false;
+        }
+    }
 
-            return intval($json_output->circledByCount);
+    /**
+     * Returns the number of Instagram followers.
+     *
+     * @param string $url the Instagram to check
+     *
+     * @return int|bool the number of +1s, or false on failure
+     */
+    private function getInstagramFollowers($url) {
+        try {
+            $json = @file_get_contents($url . '?__a=1');
+            if ($json) {
+                $json = json_decode($json);
+                if (isset($json->user->followed_by->count)) {
+                    return intval($json->user->followed_by->count);
+                }
+            }
+            return false;
         } catch (Exception $e) {
             return false;
         }
@@ -157,7 +182,7 @@ class SocialMediaUpdater extends AbstractHelper {
      *
      * @param string $url the Twitter URL to check
      *
-     * @return int the number of subscribers, or false on failure
+     * @return int|bool the number of subscribers, or false on failure
      */
     private function getTwitterFollower($url) {
         try {
@@ -169,14 +194,14 @@ class SocialMediaUpdater extends AbstractHelper {
                 'consumer_secret' => TWITTER_API_CONSUMER_SECRET
             );
             $twitterAPI = new TwitterAPIExchange($settings);
-            $follow_count = $twitterAPI->setGetfield('?screen_name=' . $name)
-                                       ->buildOauth('https://api.twitter.com/1.1/users/show.json', 'GET')
-                                       ->performRequest();
-            $data = json_decode($follow_count, true);
-            if (!$data || !isset($data['followers_count']) ) {
-                return false;
+            $json = $twitterAPI->setGetfield('?screen_name=' . $name)
+                               ->buildOauth('https://api.twitter.com/1.1/users/show.json', 'GET')
+                               ->performRequest();
+            $json = json_decode($json);
+            if (isset($json->followers_count) ) {
+                return intval($json->followers_count);
             }
-            return $data['followers_count'];
+            return false;
         } catch (Exception $e) {
             return false;
         }
@@ -187,11 +212,12 @@ class SocialMediaUpdater extends AbstractHelper {
      *
      * @param string $url the YouTube URL to check
      *
-     * @return int the number of subscribers, or false on failure
+     * @return int|bool the number of subscribers, or false on failure
      */
     private function getYoutubeSubscribers($url) {
         try {
             $username = substr($url, 24);
+            $channelId = false;
             if ( SocialMediaUpdater::startsWith($username, 'channel/') ) {
                 $temp = explode('/', $username);
                 $channelId = end($temp);
@@ -199,22 +225,27 @@ class SocialMediaUpdater extends AbstractHelper {
                 $json_url = 'https://www.googleapis.com/youtube/v3/search?part=snippet&q=' . $username
                             . '&type=channel&key=' . GOOGLE_API_KEY;
                 $json = @file_get_contents($json_url);
-                if (!$json) {
-                    return false;
+                if ($json) {
+                    $json = json_decode($json);
+                    if (isset($json->items[0]->id->channelId)) {
+                        $channelId = $json->items[0]->id->channelId;
+                    }
                 }
-                $json_output = json_decode($json);
-                $channelId = $json_output->items[0]->id->channelId;
             }
 
-            $json_url = 'https://www.googleapis.com/youtube/v3/channels?part=statistics&id=' . $channelId
-                        . '&key=' . GOOGLE_API_KEY;
-            $json = @file_get_contents($json_url);
-            if (!$json) {
-                return false;
+            if ($channelId) {
+                $json_url = 'https://www.googleapis.com/youtube/v3/channels?part=statistics&id=' . $channelId
+                            . '&key=' . GOOGLE_API_KEY;
+                $json = @file_get_contents($json_url);
+                if ($json) {
+                    $json = json_decode($json);
+                    if (isset($json->items[0]->statistics->subscriberCount)) {
+                        return intval($json->items[0]->statistics->subscriberCount);
+                    }
+                }
             }
-            $json_output = json_decode($json);
 
-            return intval($json_output->items[0]->statistics->subscriberCount);
+            return false;
         } catch (Exception $e) {
             return false;
         }