+++ CHANGELOG.md
@@ -6,12 +6,15 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 
-## Version 3.1 (unreleased)
+## Version 3.1 (2017-11-05)
 
 ### Added
+* Comparison: cron for updating the number of Instagram followers (via Instagram API)
 * Redirects for URLs without trailing slash
+* `cron.php` for cron job (both JSON/CSV export and social media update) 
 
 ### Fixed
+* Comparison: additional checks for correctness of JSON returned by APIs
 * I18n: fixed broken translations
 * Development: no redirection to HTTPS on localhost